test_recover + test_sign_in
